requirements.txt
Click==7.0
psycopg2-binary==2.7.5
